+++ src/test-support/setup-env.ts
@@ -14,37 +14,68 @@ Object.defineProperty(HTMLElement.prototype, 'clientHeight', {
   },
 });
 
-Object.defineProperty(window, 'addEventListener', {
-  configurable: true,
-  value(event, callback) {
-    if (!this.callbacks) {
-      // noinspection JSUnusedGlobalSymbols
-      this.__callbacks = {};
-    }
-    if (!this.__callbacks[event]) {
-      this.__callbacks[event] = [];
-    }
-    this.__callbacks[event].push(callback);
-  },
-});
+const eventTargets = [window, HTMLElement.prototype];
 
-Object.defineProperty(window, 'removeEventListener', {
-  configurable: true,
-  value(event, callback) {
-    if (!this.callbacks) {
+eventTargets.forEach(target => {
+  const originalAddEventListener = target.addEventListener;
+  const originalRemoveEventListener = target.removeEventListener;
+
+  Object.defineProperty(target, 'addEventListener', {
+    configurable: true,
+    value(event, callback) {
+      originalAddEventListener.call(this, event, callback);
+      if (!this.__eventListeners) {
+        // noinspection JSUnusedGlobalSymbols
+        this.__eventListeners = {};
+      }
+      if (!this.__eventListeners[event]) {
+        this.__eventListeners[event] = [];
+      }
+      this.__eventListeners[event].push(callback);
+    },
+  });
+
+  Object.defineProperty(target, 'removeEventListener', {
+    configurable: true,
+    value(event, callback) {
+      originalRemoveEventListener.call(this, event, callback);
+      if (!this.__eventListeners) {
+        // noinspection JSUnusedGlobalSymbols
+        this.__eventListeners = {};
+      }
+      if (!this.__eventListeners[event]) {
+        this.__eventListeners[event] = [];
+      }
+      this.__eventListeners[event] = this.__eventListeners[event].filter(
+        element => element !== callback,
+      );
+    },
+  });
+
+  Object.defineProperty(target, 'clearEventListeners', {
+    configurable: true,
+    value() {
       // noinspection JSUnusedGlobalSymbols
-      this.__callbacks = {};
-    }
-    if (!this.__callbacks[event]) {
-      this.__callbacks[event] = [];
-    }
-    this.__callbacks[event] = this.__callbacks[event].filter(
-      element => element !== callback,
-    );
-  },
+      this.__eventListeners = {};
+    },
+  });
 });
 
 Object.defineProperty(HTMLMediaElement.prototype, 'load', {
   configurable: true,
   value: jest.fn(),
+});
+
+Object.defineProperty(HTMLElement.prototype, 'setBoundingClientRect', {
+  configurable: true,
+  value: (domRect: DOMRect) => {
+    this.__boundingClientRect = domRect;
+  },
+});
+
+Object.defineProperty(HTMLElement.prototype, 'getBoundingClientRect', {
+  configurable: true,
+  value: () => {
+    return this.__boundingClientRect;
+  },
 });