+++ src/test-support/setup.ts
@@ -1,6 +1,7 @@
 import axios from 'axios';
 import AxiosLogger from 'axios-logger';
 import MockFetchVerify from './MockFetchVerify';
+import { HasEventListeners } from './types';
 
 beforeEach(() => {
   (axios.interceptors.request as any).handlers = [];
@@ -10,4 +11,5 @@ beforeEach(() => {
 
   // Can you believe JEST doesn't automatically clean up?
   document.body.innerHTML = '';
+  ((window as any) as HasEventListeners).clearEventListeners();
 });