+++ src/utils/convertPlaybackResource.ts
@@ -1,6 +1,11 @@
 import { parse, toSeconds } from 'iso8601-duration';
 import { Link } from '../types/Link';
-import { Playback, StreamPlayback, YoutubePlayback } from '../types/Playback';
+import {
+  isStreamPlayback,
+  Playback,
+  StreamPlayback,
+  YoutubePlayback,
+} from '../types/Playback';
 
 const convertPlaybackResource = (
   rawPlayback,
@@ -24,6 +29,23 @@ const convertPlaybackResource = (
 
   if (rawPlayback.type === 'STREAM') {
     (playback as StreamPlayback).streamUrl = rawPlayback.streamUrl;
+  }
+
+  if (rawPlayback._links.thumbnailApi) {
+    playback.links.thumbnailApi = new Link(rawPlayback._links.thumbnailApi);
+  } else if (
+    process.env.NODE_ENV === 'development' &&
+    isStreamPlayback(playback)
+  ) {
+    const kalturaId = playback.thumbnailUrl.replace(
+      /.*\/entry_id\/([^\/]*)\/.*/,
+      '$1',
+    );
+
+    playback.links.thumbnailApi = new Link({
+      href: `https://cdnapisec.kaltura.com/p/1776261/thumbnail/entry_id/${kalturaId}/width/{thumbnailWidth}/vid_slices/{videoSlices}`,
+      templated: true,
+    });
   }
 
   return playback;