+++ src/utils/index.ts
@@ -1,1 +1,3 @@
 export const noop: (args?: any) => any = () => {};
+
+export const withPx = (value: string | number) => `${value}px`;